Ade cart controller

+++ src/main/java/com/codeblue/montreISTA/service/implement/CartServiceImpl.java
@@ -5,6 +5,7 @@ import com.codeblue.montreISTA.entity.*;
 import com.codeblue.montreISTA.repository.*;
 import com.codeblue.montreISTA.service.CartService;
 import lombok.AllArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import java.util.ArrayList;
 import java.util.List;
@@ -35,7 +36,7 @@ public class CartServiceImpl implements CartService {
 
     @Override
     public List<CartResponseDTO> findByBuyer(String keyword) throws Exception {
-        List<Cart> results = this.cartRepository.findByBuyerUserNameIgnoreCaseContainingOrderByCartIdAsc(keyword);
+        List<Cart> results = this.cartRepository.findByBuyerUserUsernameIgnoreCaseContainingOrderByCartIdAsc(keyword);
         if(results.isEmpty()){
             throw new Exception("Carts not found");
         }
@@ -44,7 +45,7 @@ public class CartServiceImpl implements CartService {
 
     @Override
     public List<CartResponseDTO> findBySeller(String keyword) throws Exception {
-        List<Cart> results = this.cartRepository.findByProductSellerUserIdNameIgnoreCaseContainingOrderByCartIdAsc(keyword);
+        List<Cart> results = this.cartRepository.findByProductSellerUserIdUsernameIgnoreCaseContainingOrderByCartIdAsc(keyword);
         if(results.isEmpty()){
             throw new Exception("Carts not found");
         }
@@ -70,14 +71,14 @@ public class CartServiceImpl implements CartService {
     }
 
     @Override
-    public CartResponseDTO createCart(CartRequestDTO cartRequestDTO) throws Exception {
-        Optional<Order> orderBuyerId = orderRepository.findFirstByListCartBuyerBuyerIdOrderByModifiedAtDesc(cartRequestDTO.getBuyer_id());
+    public CartResponseDTO createCart(CartRequestDTO cartRequestDTO, Authentication authentication) throws Exception {
+        Optional<Order> orderBuyerId = orderRepository.findFirstByListCartBuyerUserUsernameOrderByModifiedAtDesc(authentication.getName());
         Product productId = productRepository.findById(cartRequestDTO.getProduct_id()).orElseThrow(()->new Exception("Product not Found"));
-        Buyer buyer = buyerRepository.findById(cartRequestDTO.getBuyer_id()).orElseThrow(()->new Exception("Buyer not Found"));
-        Long orderId;
-        if(buyer.getUser().getUserId().equals(productId.getSeller().getUserId().getUserId())){
+        Buyer buyer = buyerRepository.findByUserUsername(authentication.getName()).orElseThrow(()->new Exception("Buyer not Found"));
+        if(buyer.getUser().equals(productId.getSeller().getUserId())){
             throw new Exception("You can't order your own product honey");
         }
+        Long orderId;
         if(orderBuyerId.isPresent()){
             orderId = orderBuyerId.get().getOrderId();
         }else {
@@ -92,7 +93,7 @@ public class CartServiceImpl implements CartService {
             Order saveOrder = orderRepository.save(newOrder);
             orderId = saveOrder.getOrderId();
         }
-        Cart saveCart = this.requestToEntity(cartRequestDTO, orderId);
+        Cart saveCart = this.requestToEntity(cartRequestDTO, orderId, buyer);
 
         //update Price
         this.updatePrice(orderId);
@@ -103,18 +104,17 @@ public class CartServiceImpl implements CartService {
     }
 
     @Override
-    public List<CartResponseDTO> wishlistToCart(Long id) throws Exception {
-        List<Wishlist> wishlists = wishlistRepository.findByBuyerBuyerIdOrderByModifiedAtDesc(id);
+    public List<CartResponseDTO> wishlistToCart(Authentication authentication) throws Exception {
+        List<Wishlist> wishlists = wishlistRepository.findByBuyerUserUsername(authentication.getName());
         if(wishlists.isEmpty()){
             throw new Exception("Your wishlist is empty");
         }
         List<CartResponseDTO> carts = new ArrayList<>();
         for(Wishlist wishlist:wishlists){
            CartRequestDTO cartRequestDTO = new CartRequestDTO();
-           cartRequestDTO.setBuyer_id(wishlist.getBuyer().getBuyerId());
            cartRequestDTO.setProduct_id(wishlist.getProduct().getProductId());
            cartRequestDTO.setQuantity(wishlist.getQuantity());
-           CartResponseDTO cartResponseDTO = this.createCart(cartRequestDTO);
+           CartResponseDTO cartResponseDTO = this.createCart(cartRequestDTO,authentication);
            carts.add(cartResponseDTO);
         }
         if(carts.isEmpty()){
@@ -126,23 +126,34 @@ public class CartServiceImpl implements CartService {
     }
 
     @Override
-    public CartResponseDTO updateCart(CartRequestDTO cartRequestDTO, Long id) throws Exception {
-        Cart cart = this.cartRepository.findById(id).orElseThrow(()->new Exception("Cart not found"));
+    public CartResponseDTO updateCart(CartRequestDTO cartRequestDTO, Long id,Authentication authentication) throws Exception {
+        Cart cart = cartRepository.findById(id).orElseThrow(()->new Exception("Cart not found"));
+        Buyer buyer = buyerRepository.findByUserUsername(authentication.getName()).orElseThrow(()->new Exception("Buyer not Found"));
         Long orderId = cart.getOrder().getOrderId();
-        Cart saveCart = this.requestToEntity(cartRequestDTO,orderId);
+        if(!buyer.equals(cart.getBuyer())){
+            throw new Exception("You can't update other cart");
+        }
+        if(buyer.getUser().equals(cart.getProduct().getSeller().getUserId())){
+            throw new Exception("You can't order your own product honey");
+        }
+        Cart saveCart = this.requestToEntity(cartRequestDTO,orderId,buyer);
         saveCart.setCartId(id);
         saveCart.setQuantity(cartRequestDTO.getQuantity());
         //update Price
         this.updatePrice(orderId);
-        Cart cartResponse = this.cartRepository.save(saveCart);
+        Cart cartResponse = cartRepository.save(saveCart);
         return convertDTO(cartResponse);
     }
 
     @Override
-    public void deleteById(Long id) throws Exception {
+    public void deleteById(Long id,Authentication authentication) throws Exception {
         Optional<Cart> cartId = this.cartRepository.findById(id);
-        if(cartId.isEmpty()){
+        Buyer buyer = buyerRepository.findByUserUsername(authentication.getName()).orElseThrow(()->new Exception("Buyer not Found"));
+         if(cartId.isEmpty()){
             throw new Exception("Cart not found");
+        }
+        if(!buyer.equals(cartId.get().getBuyer())){
+            throw new Exception("You can't delete other cart");
         }
         this.cartRepository.deleteById(id);
     }
@@ -164,16 +175,9 @@ public class CartServiceImpl implements CartService {
         return cart.convertToResponse(photosDTO, categoriesDTO);
     }
 
-    public Cart requestToEntity (CartRequestDTO cartRequestDTO, Long orderId)throws Exception{
-        Optional<Product> productById = productRepository.findById(cartRequestDTO.getProduct_id());
-        Optional<Order> orderById = orderRepository.findById(orderId);
-        Optional<Buyer> buyerById = buyerRepository.findById(cartRequestDTO.getBuyer_id());
-        if(productById.isEmpty() || orderById.isEmpty() || buyerById.isEmpty()){
-            throw new Exception("Data not found");
-        }
-        Product product = productById.get();
-        Order order = orderById.get();
-        Buyer buyer = buyerById.get();
+    public Cart requestToEntity (CartRequestDTO cartRequestDTO, Long orderId, Buyer buyer)throws Exception{
+        Product product = productRepository.findById(cartRequestDTO.getProduct_id()).orElseThrow(()->new Exception("Product not found"));
+        Order order = orderRepository.findById(orderId).orElseThrow(()->new Exception("Order not found"));
         return cartRequestDTO.convertToEntity(buyer,product,order);
     }
